papertrail now more generic

+++ logspout/recipes/install.rb
@@ -1,34 +1,40 @@
 Chef::Log.info("Setting up papertrail")
 
-Chef::Log.info
+node[:deploy].each do |application, deploy|
 
-hostname = "#{node[:opsworks][:stack][:name]}-#{node[:opsworks][:instance][:hostname]}"
+    if node[:opsworks][:instance][:layers].first != deploy[:environment_variables][:layer]
+        Chef::Log.warn("Skipping the deployment of this nginx because it does not match the app layer")
+        next
+    end
 
-bash "docker-logspout-cleanup" do
-    user "root"
-    code <<-EOH
-        if docker ps | grep logspout;
-        then
-            docker stop logspout
-            sleep 3
-            docker rm -f logspout
-        fi
-        if docker ps -a | grep logspout;
-        then
-            docker rm -f logspout
-        fi
-        if docker images | grep progrium/logspout;
-        then
-            docker rmi -f $(docker images | grep -m 1 progrium/logspout | awk {'print $3'})
-        fi
-    EOH
+    hostname = "#{node[:opsworks][:stack][:name]}-#{node[:opsworks][:instance][:hostname]}"
+
+    bash "docker-logspout-cleanup" do
+        user "root"
+        code <<-EOH
+            if docker ps | grep logspout;
+            then
+                docker stop logspout
+                sleep 3
+                docker rm -f logspout
+            fi
+            if docker ps -a | grep logspout;
+            then
+                docker rm -f logspout
+            fi
+            if docker images | grep progrium/logspout;
+            then
+                docker rmi -f $(docker images | grep -m 1 progrium/logspout | awk {'print $3'})
+            fi
+        EOH
+    end
+
+    Chef::Log.info('docker-logspout start')
+    bash "docker-logspout" do
+        user "root"
+        code <<-EOH
+            docker run -d -h #{hostname} --name logspout -v /var/run/docker.sock:/tmp/docker.sock progrium/logspout syslog://#{deploy[:environment_variables][:PAPERTRAIL_URL]}:#{deploy[:environment_variables][:PAPERTRAIL_PORT]}
+        EOH
+    end
+    Chef::Log.info('docker-logspout stop')
 end
-
-Chef::Log.info('docker-logspout start')
-bash "docker-logspout" do
-    user "root"
-    code <<-EOH
-        docker run -d -h #{hostname} --name logspout -v /var/run/docker.sock:/tmp/docker.sock progrium/logspout syslog://logs2.papertrailapp.com:44126
-    EOH
-end
-Chef::Log.info('docker-logspout stop')
